requirements.txt
torch
transformers
pandas
Pillow
torchdata
numpy
gcsfs